:truck: 방 설정 컴포넌트 분리 #69
