+++ src/pages/play/lobby/types.d.ts
@@ -1,5 +1,3 @@
-import * as constants from './constants';
-
 type playerColorType =
-  | (typeof constants.PLAYER_COLORS)[number]
-  | typeof constants.EMPTY_PLAYER_COLOR;
+  | (typeof import('./constants').PLAYER_COLORS)[number]
+  | typeof import('./constants').EMPTY_PLAYER_COLOR;
